fix(persistence.repository): Move controllers to 'controllers' package
