+++ persistence/repository/src/main/java/org/company/persistence/web/controller/ArticleController.java
@@ -1,4 +1,4 @@
-package org.company.persistence.web;
+package org.company.persistence.web.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
